# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SWARM Rendezvous API",
    "description": "Rendezvous service for the SWARM P2P media suite: accounts, swarms, join codes, device registry, and WSS signaling (see /api/v1/ws — not representable in OpenAPI).",
    "license": {
      "name": "GPL-3.0-or-later",
      "identifier": "GPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/devices/register": {
      "post": {
        "tags": [
          "devices"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterDeviceResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/{device_id}/metadata": {
      "patch": {
        "tags": [
          "devices"
        ],
        "operationId": "patch_metadata",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/swarms/join": {
      "post": {
        "tags": [
          "devices"
        ],
        "operationId": "join_swarm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinSwarmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SwarmSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "Uniform error body for every non-2xx REST response.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "DeviceRegistration": {
        "type": "object",
        "required": [
          "name",
          "device_type",
          "machine_id",
          "cert_fingerprint",
          "platform",
          "app_version"
        ],
        "properties": {
          "app_version": {
            "type": "string"
          },
          "cert_fingerprint": {
            "type": "string",
            "description": "Lowercase hex SHA-256 of the device's self-signed certificate (DER)."
          },
          "device_type": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "machine_id": {
            "type": "string",
            "description": "Stable machine identity (persisted NIC-MAC-derived id), so\nre-registration updates the existing device row instead of duplicating."
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary user-visible key/value metadata (hostname, cpu, disk, ...).",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "platform": {
            "type": "string",
            "description": "e.g. \"macos-aarch64\", \"firetv-api25\"."
          }
        },
        "additionalProperties": false
      },
      "DeviceType": {
        "type": "string",
        "enum": [
          "client",
          "server",
          "both"
        ]
      },
...
      "JoinSwarmRequest": {
        "type": "object",
        "description": "`POST /api/v1/swarms/join` — add an already-registered device (Bearer auth)\nto another swarm with a fresh join code.",
        "required": [
          "code"
        ],
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "MetadataPatchRequest": {
        "type": "object",
        "description": "`PATCH /api/v1/devices/{device_id}/metadata` — a device updating its own\narbitrary key/value metadata (e.g. a server self-reporting the address\npeers should dial: `{\"peer_addr\": \"192.168.1.50:8543\"}`).",
        "required": [
          "metadata"
        ],
        "properties": {
          "metadata": {
            "type": "object",
            "description": "Keys set to an empty string are removed; others upserted.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
...
      "RegisterDeviceRequest": {
        "type": "object",
        "description": "`POST /api/v1/devices/register` — redeem a single-use, swarm-scoped join\ncode. The cert fingerprint submitted here is the TOFU trust anchor for all\nlater peer connections.",
        "required": [
          "code",
          "device"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "8-digit join code generated in the STUN web UI, scoped to one swarm."
          },
          "device": {
            "$ref": "#/components/schemas/DeviceRegistration"
          }
        },
        "additionalProperties": false
      },
      "RegisterDeviceResponse": {
        "type": "object",
        "required": [
          "access_token",
          "device_id",
          "swarm"
        ],
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Opaque bearer token; store encrypted at rest on-device."
          },
          "device_id": {
            "type": "string"
          },
          "swarm": {
            "$ref": "#/components/schemas/SwarmSummary"
          }
        }
      },
...
      "StatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
...
      "SwarmSummary": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "swarm_session"
      }
    }
  },
...
}